fix(cli): fix `completion` and `make-this-awesome` command invalid problem(#1889)
